rag_md_bot/requirements.txt
sentence-transformers==2.7.0
numpy
tqdm
python-dotenv==1.0.1
chromadb
ollama
